Refactor usePyishObjectMapper into LegacyOverrides
