+++ src/main/java/com/hubspot/jinjava/JinjavaConfig.java
@@ -51,7 +51,6 @@ public class JinjavaConfig {
   private final Map<Context.Library, Set<String>> disabled;
   private final boolean failOnUnknownTokens;
   private final boolean nestedInterpretationEnabled;
-  private final boolean usePyishObjectMapper;
   private final RandomNumberGeneratorStrategy randomNumberGenerator;
   private final boolean validationMode;
   private final long maxStringLength;
@@ -103,7 +102,6 @@ public class JinjavaConfig {
     failOnUnknownTokens = builder.failOnUnknownTokens;
     maxOutputSize = builder.maxOutputSize;
     nestedInterpretationEnabled = builder.nestedInterpretationEnabled;
-    usePyishObjectMapper = builder.usePyishObjectMapper;
     randomNumberGenerator = builder.randomNumberGeneratorStrategy;
     validationMode = builder.validationMode;
     maxStringLength = builder.maxStringLength;
@@ -176,10 +174,6 @@ public class JinjavaConfig {
     return nestedInterpretationEnabled;
   }
 
-  public boolean isUsePyishObjectMapper() {
-    return usePyishObjectMapper;
-  }
-
   public boolean isValidationMode() {
     return validationMode;
   }
@@ -235,7 +229,6 @@ public class JinjavaConfig {
     private int maxMacroRecursionDepth;
     private boolean failOnUnknownTokens;
     private boolean nestedInterpretationEnabled = true;
-    private boolean usePyishObjectMapper = false;
     private RandomNumberGeneratorStrategy randomNumberGeneratorStrategy =
       RandomNumberGeneratorStrategy.THREAD_LOCAL;
     private boolean validationMode = false;
@@ -330,11 +323,6 @@ public class JinjavaConfig {
       return this;
     }
 
-    public Builder withUsePyishObjectMapper(boolean usePyishObjectMapper) {
-      this.usePyishObjectMapper = usePyishObjectMapper;
-      return this;
-    }
-
     public Builder withValidationMode(boolean validationMode) {
       this.validationMode = validationMode;
       return this;
@@ -366,7 +354,7 @@ public class JinjavaConfig {
     }
 
     /**
-     * @deprecated  Replaced by {@link LegacyOverrides.Builder#withIterateOverMapKeys(boolean)} ()}
+     * @deprecated  Replaced by {@link LegacyOverrides.Builder#withIterateOverMapKeys(boolean)}}
      */
     @Deprecated
     public Builder withIterateOverMapKeys(boolean iterateOverMapKeys) {
