+++ src/test/java/com/hubspot/jinjava/BaseJinjavaTest.java
@@ -8,6 +8,13 @@ public abstract class BaseJinjavaTest {
   @Before
   public void baseSetup() {
     jinjava =
-      new Jinjava(JinjavaConfig.newBuilder().withUsePyishObjectMapper(true).build());
+      new Jinjava(
+        JinjavaConfig
+          .newBuilder()
+          .withLegacyOverrides(
+            LegacyOverrides.newBuilder().withUsePyishObjectMapper(true).build()
+          )
+          .build()
+      );
   }
 }
